+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# staged
+
+- FIX: allow setting custom http scheme
+
 # v2.4.2
 
 - Dependancy updates